requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
pydantic[email]